working towards automation

+++ scripts/test_lustre.sh
@@ -1,29 +1,21 @@
 #!/bin/bash
 
+LLMOUNT="./lustre/tests/llmount.sh"
+[ ! -f "$LLMOUNT_SH" ] && { echo "Warning: $LLMOUNT_SH does not exist."; exit 1; }
+[ "$(id -u)" -ne 0 ] && { echo "Warning: This script must be run as root."; exit 1; }
+
 testfile=/mnt/lustre/file.out
 
-# Get the primary IP address
-IP_ADDRESS=$(hostname -I)
-
-# Set the hostname 
-# not necessary anymore since the VM creation tools handle this
-# hname=mylustrehost
-# hostnamectl set-hostname $hname 
-
-# also update /etc/hosts with this new hostname
-# hmmm, I wonder if this is necessary?
-echo "No longer setting the hostname in /etc/hosts, is that a problem?"
-# echo "$IP_ADDRESS $hname" >> /etc/hosts
-
 # first clean it up in case we forgot to do that
-./lustre/tests/llmountcleanup.sh
 \rm -rf /mnt/lustre
 
-FSTYPE=zfs ./lustre/tests/llmount.sh
+FSTYPE=zfs $LLMOUNT 
 
+# test the default striping behavior
 df -h /mnt/lustre
 ./lustre/utils/lfs getstripe /mnt/lustre/
 dd if=/dev/zero of=$testfile bs=1MB count=400
+echo "The default striping pattern"
 df -h /mnt/lustre-*
 \rm $testfile
 
@@ -33,5 +25,6 @@ df -h /mnt/lustre-*
 ./lustre/utils/lfs getstripe /mnt/lustre/
 
 dd if=/dev/zero of=$testfile bs=1MB count=400
+echo "The modified striping pattern"
 df -h /mnt/lustre-*
 \rm $testfile
